+++ src/main/java/com/malbano/organizai/transacoes/entity/TipoTransacao.java
@@ -9,7 +9,7 @@ public class TipoTransacao {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "tipo_transacao_id")
-    private Integer tipoTransacaoId;
+    private Long tipoTransacaoId;
 
     @Column(name = "descricao_transacao", nullable = false, length = 20)
     private String descricaoTransacao;
